critical fix(history route) fixed issue with displaying history route
Fixed issue stopping history route from working
this is due to inconsistency in the app of sell and buy being also named purchase and sale.
In the app, (apart from the message to the user where appropriate) it should use buy and sell.
I have already begun the process of moving everything over to buy and sell